requirements_old.txt
requests
beautifulsoup4
lxml
python-dotenv
google-genai
pydantic
pandas
google-api-core
